research/corruption-investigation/briefs/epstein-doj-analysis/requirements.txt
# Core PDF processing
pypdf>=3.17.0
pdfplumber>=0.10.3
pdf2image>=1.16.3

# OCR
pytesseract>=0.3.10

# NLP and NER
spacy>=3.7.0
transformers>=4.36.0
torch>=2.1.0

# Hugging Face
datasets>=2.16.0
huggingface-hub>=0.20.0

# Network analysis (for network_mapper.py)
networkx>=3.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
pandas>=2.1.0
psutil>=5.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
reportlab>=4.0.0  # For generating test PDFs
